requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib
bcrypt
python-dotenv
openai
jinja2